more repairs and post install hook clean dup

+++ scripts/postinstall.sh
@@ -10,9 +10,25 @@ for project in */ ; do
   p=${project::-1}
   rm -rf node_modules
   ln -s ../../node_modules
-  rm ../../node_modules/@lrnwebcomponents/${p}
-  mkdir ../../node_modules/@lrnwebcomponents/${p}
-  ln -s ../../../elements/${p}/${p}.js ../../node_modules/@lrnwebcomponents/${p}/${p}.js
-  ln -s ../../../elements/${p}/lib ../../node_modules/@lrnwebcomponents/${p}/lib
+  DIRECTORY="../../node_modules/@lrnwebcomponents/${p}"
+  if [ -d "$DIRECTORY" ]; then
+    rm ../../node_modules/@lrnwebcomponents/${p}
+    mkdir ../../node_modules/@lrnwebcomponents/${p}
+    if [ -f "${p}.js" ]; then
+      ln -s ../../../elements/${p}/${p}.js ../../node_modules/@lrnwebcomponents/${p}/${p}.js
+    fi
+    if [ -d "lib" ]; then
+      ln -s ../../../elements/${p}/lib ../../node_modules/@lrnwebcomponents/${p}/lib
+    fi
+    if [ -d "build" ]; then
+      ln -s ../../../elements/${p}/build ../../node_modules/@lrnwebcomponents/${p}/build
+    fi
+    if [ -d "src" ]; then
+      ln -s ../../../elements/${p}/src ../../node_modules/@lrnwebcomponents/${p}/src
+    fi
+    if [ -d "dist" ]; then
+      ln -s ../../../elements/${p}/dist ../../node_modules/@lrnwebcomponents/${p}/dist
+    fi
+  fi
   cd ../
 done
